+++ NEWS.md
@@ -6,6 +6,7 @@
 # mlr3torch 0.1.2
 
 * Don't use deprecated `data_formats` anymore
+* Added `CallbackSetTB`, which allows logging that can be viewed by TensorBoard.
 
 # mlr3torch 0.1.1